# Do not mutate IEnumerable and null values when formatting