--- a/src/libraries/Microsoft.Extensions.Logging/tests/Common/FormattedLogValuesTest.cs
+++ b/src/libraries/Microsoft.Extensions.Logging/tests/Common/FormattedLogValuesTest.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using Xunit;
 
@@ -92,6 +93,29 @@ namespace Microsoft.Extensions.Logging.Test
                 var logValues = new FormattedLogValues(format, new object[] { "arg1" });
                 logValues.ToString();
             });
+        }
+
+        [Fact]
+        public void LogValues_WithNullAndEnumerable_IsNotMutatingParameter()
+        {
+            var format = "TestMessage {Param1} {Param2} {Param3} {Param4}";
+            var param1 = 1;
+            var param2 = (string)null;
+            var param3 = new[] { 1, 2, 3, 4 };
+            var param4 = "string";
+
+            var logValues = new FormattedLogValues(format, param1, param2, param3, param4);
+            logValues.ToString();
+
+            var state = logValues.ToArray();
+            Assert.Equal(new[]
+            {
+                new KeyValuePair<string, object>("Param1", param1),
+                new KeyValuePair<string, object>("Param2", param2),
+                new KeyValuePair<string, object>("Param3", param3),
+                new KeyValuePair<string, object>("Param4", param4),
+                new KeyValuePair<string, object>("{OriginalFormat}", format),
+            }, state);
         }
 
         [Fact]