This fixes https://github.com/dotnet/runtime/issues/36165, https://github.com/dotnet/runtime/issues/36025, and https://github.com/dotnet/runtime/issues/48873. 

I tossed my first fix I did that checked the array reference on every iteration in the loop and used the snippet @stephentoub proposed in the following discussion that was more elegant: https://github.com/dotnet/runtime/pull/38734

Not sure the FormatQuick (naming is hard!) is really needed (a few more bytes allocated according to my tests), but the LoggerMessage is supposed to be used for high-performance logging scenarios, so maybe worth it anyways.

I created a pull-request in benchmarks for related microbenchmarks: https://github.com/dotnet/performance/pull/1708