@pergardebrink 2021-03-07
The JsonConsoleFormatter (and any logging provider using the state) will have a breaking change now that the state is passed to the log provider exactly as provided, without being mutated (which, in my opinion, was a bug before and was the reason I filed https://github.com/dotnet/runtime/issues/48873). 

Does this need to be added to any documentation/release notes if this pull-request is accepted?