@tarekgh 2021-03-22
@pergardebrink the change looks good. Thanks for adding the benchmark for logging formatting too. could you please paste here some perf numbers before and after your changes showing the memory allocations?

CC @stephentoub @davidfowl