@pergardebrink 2021-03-27
> @pergardebrink the change looks good. Thanks for adding the benchmark for logging formatting too. could you please paste here some perf numbers before and after your changes showing the memory allocations?

Is there any special scenarios you'd like to see? The benchmark I added to the performance repo for the case with enumerable parameter, along with existing ones look like the following, but I can run some extra variations on number of arguments. How many arguments to a log message would be interesting to see. In our codebase at our company, I'd say most log messages have on average 2-3, arguments up to a maximum of 10?:

Before:
|                           Method |      Mean |     Error |    StdDev |    Median |       Min |       Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |----------:|----------:|----------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
|       TwoArguments_DefineMessage | 161.13 ns |  1.679 ns |  1.571 ns | 161.15 ns | 158.29 ns | 164.28 ns | 0.0090 |     - |     - |      80 B |
|      FourArguments_DefineMessage | 268.59 ns |  2.042 ns |  1.910 ns | 269.19 ns | 264.81 ns | 271.03 ns | 0.0207 |     - |     - |     176 B |
|                      NoArguments |  26.89 ns |  0.238 ns |  0.222 ns |  26.84 ns |  26.51 ns |  27.35 ns |      - |     - |     - |         - |
|                     TwoArguments | 201.28 ns |  2.284 ns |  2.243 ns | 201.52 ns | 197.89 ns | 206.11 ns | 0.0139 |     - |     - |     120 B |
| FourArguments_EnumerableArgument | 875.65 ns | 10.929 ns | 10.223 ns | 879.34 ns | 850.23 ns | 885.08 ns | 0.0702 |     - |     - |     600 B |


After:
|                           Method |      Mean |    Error |   StdDev |    Median |       Min |       Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |----------:|---------:|---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
|       TwoArguments_DefineMessage | 160.32 ns | 1.423 ns | 1.331 ns | 160.58 ns | 157.56 ns | 162.81 ns | 0.0091 |     - |     - |      80 B |
|      FourArguments_DefineMessage | 268.60 ns | 2.429 ns | 2.272 ns | 268.77 ns | 263.66 ns | 272.80 ns | 0.0208 |     - |     - |     176 B |
|                      NoArguments |  27.28 ns | 0.549 ns | 0.564 ns |  27.36 ns |  26.49 ns |  28.17 ns |      - |     - |     - |         - |
|                     TwoArguments | 196.46 ns | 2.386 ns | 2.232 ns | 196.56 ns | 193.16 ns | 200.26 ns | 0.0142 |     - |     - |     120 B |
| FourArguments_EnumerableArgument | 895.14 ns | 6.801 ns | 6.362 ns | 895.47 ns | 886.72 ns | 908.54 ns | 0.0783 |     - |     - |     656 B |