@pergardebrink 2021-03-27
Added local benchmark for 10 arguments with/without an enumerable also (the enumerable item came in position 5 in the array)

Before:
|                           Method |        Mean |     Error |    StdDev |      Median |         Min |         Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |------------:|----------:|----------:|------------:|------------:|------------:|-------:|------:|------:|----------:|
|  TenArguments_EnumerableArgument | 1,187.31 ns | 25.175 ns | 28.992 ns | 1,195.83 ns | 1,130.22 ns | 1,235.98 ns | 0.0888 |     - |     - |     768 B |
| TenArguments_NoEnumerableArgument |   624.23 ns |  4.705 ns |  4.401 ns |   622.92 ns |   618.17 ns |   632.00 ns | 0.0388 |     - |     - |     344 B |

After:
|                           Method |        Mean |     Error |    StdDev |      Median |         Min |         Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |------------:|----------:|----------:|------------:|------------:|------------:|-------:|------:|------:|----------:|
|  TenArguments_EnumerableArgument | 1,201.41 ns |  4.526 ns |  4.013 ns | 1,201.59 ns | 1,194.70 ns | 1,208.46 ns | 0.1001 |     - |     - |     872 B |
| TenArguments_NoEnumerableArgument |   620.78 ns | 18.114 ns | 20.860 ns |   616.15 ns |   598.21 ns |   669.17 ns | 0.0409 |     - |     - |     344 B |