@tarekgh 2021-03-28
@pergardebrink thanks a lot for updating the code and getting the perf numbers. 

@davidfowl @stephentoub there is some extra allocations happening but this to fix the original concern. I am seeing the result is reasonable. do you have any feedback before we go ahead and merge this PR?