@davidfowl 2021-03-28
As long as the extra allocations are when there's an enumerable then that's fine